+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+###### 4.06.2024
+
+> ESP32 admin delete all data status functionality
+> ESP32 admin view all verified users functionality
+
 ###### 4.05.2024
 
 > ESP32 completely functioning source code compileation error fix //TODO: Add custom library